Initial commit of the rewritten prototype
This has started with a small prototype which used Nix as a preprocessor
to generate a header file with preinitialized rules. However a socket
wrapper that is based on rules is not only useful for NixOS or nixcloud
in particular but might be generally useful.

Now we have a small command line program which accepts a JSON file as
its first argument, so we no longer need to compile the program
specifically for our use case and can supply rules at runtime.

However, there are still things in dire need of overhaul/improvement:

  * We need a few more mutexes, because right now we only have one for
    writing to the global map of active sockets, which will quickly fall
    apart as soon as we're trying to read a value while another thread
    is writing to it.

  * Most of the tests (specifically the integration tests) are not yet
    ported over from the prototype implementation. They're very very
    slow and their implementation is pretty much quick & dirty, so we
    need to reimplement those tests to also be more specific.

  * The SocketInfo struct could also need a little bit of love in terms
    of data structures that are better suited. For our child/parent-
    relationships a more tree-like structure might fit better.

  * Split off the helper functions from preload.cc, because that file
    currently is pretty messy and mixed with code that is very rarely
    touched (eg. the "real" namespace).

  * Instead of relying on RTLD_NEXT it would make more sense to
    reference the C library directly, so that we don't accidentally wrap
    the wrong library function.

  * Come up with a good solution to assign/rotate peer addresses.

There are even more little things we can improve, but those are at least
the most pressing points to make it work with applications out there
that rely heavily on multi-threading.

Compared to the prototype implementation this implementation works with
Erlang, however I still manage to hit critical sections from time to
time (especially on high load).

Signed-off-by: aszlig <aszlig@nix.build>