Add tests for simple connections
This is implemented by having a small helper program that is executed
for both client and server connections. We need this as an external
program because we want to run it with ip2unix.

The helper program acts as client or server depending on the command
line arguments given and just implements a very simple test sending a
payload to the server, which sends it back upper-cased and the client
just checks whether it has been upper-cased correctly.

It also has different options to run either using fork or threads, so we
should be able to test whether our mutexes work as they should.

Within the actual tests, we just do a 1:1 for the client and server,
optionally running several connections at one on the client. The test
cases are actually quite simple right now and we obviously need to have
a few more of them.

Signed-off-by: aszlig <aszlig@nix.build>