UdsmapRule: Provide default values
It's not a very good idea to leave those values unintialized because
depending on the compiler this could end up with very weird behavior.

So let's initialize the whole struct, so that we don't run into future
surprises. We thus can also remove the initialization of the
socket_activation field in rules.cc.

Signed-off-by: aszlig <aszlig@nix.build>