Add test for systemd socket activation
This depends on the 'systemd-socket-activate' helper to provide a
similar environment to what we have when launched as a service unit
bound to a socket unit.

Most of the changes were plumbing so that the test helper methods allow
to pass additional commands before ip2unix and also make sure we don't
try to wait for the connector helper to be started, as systemd only
starts it with the first connection.

Signed-off-by: aszlig <aszlig@nix.build>