Allew to define placeholders for socketPath
This is especially useful if we want to have a single rule matching
several connections and also will come in handy once we fully support
binding to port 0.

With this, we now re-add the sockpath field of SocketInfo and set it to
the expanded socket path, so that we can later unlink the right socket
file.

Signed-off-by: aszlig <aszlig@nix.build>