tests/connector: Retry after failed connection
Betwen creating/binding the socket and listening there is a small time
frame where connecting to the socket will be refused.

So instead of waiting for the socket path to exist or doing anything
more fishy, just retry the connection indefinitely. The test runner will
time out after 30 seconds anyway, so we don't need to handle this.

However, for debugging I added a message to be printed after it took
longer than 10 seconds to connect.

Signed-off-by: aszlig <aszlig@nix.build>