Overhaul, add and refine all mutexes
So far we only had one mutex and it was used all over the place and
didn't really cover our critical sections.

Now, the g_sockinfo_mutex (previously g_mutex) is used *only* within the
wrapped syscalls and also *only* if we actually want to access the
g_active_sockets variable in *any* way.

The g_rules_mutex is *only* used whenever we want to access or write to
g_rules, so that there should be no way we accidentally write the rules
twice when run concurrently.

And the last mutex is just g_dlsym_mutex, which is used whenever we load
symbols from the C library.

I did 1000 runs of our test suite and while before the threading tests
were randomly failing, we now have a 100% success rate.

Signed-off-by: aszlig <aszlig@nix.build>