test_rules: Add a few type-related validity tests
We're going to get rid of the JSON schema validation (for reasons I'll
describe in a future commit), so we need to make sure that our rule
checker still works correctly and not just trust that schema validation
will take care of everything.

So for the upcoming conversion to YAML based rules, we need to have
tests in place so we don't weaken our schema checking.

Signed-off-by: aszlig <aszlig@nix.build>