Switch rule file format to YAML
First of all, I personally find YAML way more readable than JSON and
second it's a superset of JSON, so with the switch to YAML we
effectively don't lose anything in terms of compatibility (not that it
would be an issue at this stage of development).

Before this switch we used RapidJSON to do the parsing *and* validation,
which yaml-cpp doesn't do. This however is really a none-issue because
the errors RapidJSON 1.1.0 gives back are really not very useful.

So instead, I'm manually checking the individual options now, which also
helps us prepare for the upcoming work to support rules directly on the
command line.

While the checking could be implemented in a less ugly way (eg.
checking ports is currently quite stupid), it still has the benefit that
we now actually check whether the IP address given in an address
matching rule is valid by using inet_pton() and I also added a few tests
with a few valid and invalid addresses.

Signed-off-by: aszlig <aszlig@nix.build>