Pass rule contents via environment variable
My first stab at this was to use memfd_create() to pass down a file
descriptor to the executable, because we are limited in terms of process
environment and it could also get us into trouble if we have a \0 in the
rule file.

However in practice the memfd_create() method didn't work out so well,
because some programs tend to close file descriptions either directly or
indirectly so we end up in a completely inconsistent state with some
processes being correctly wrapped and some not.

So I simply went for re-encoding the whole YAML file into an environment
variable because if we *should* hit limits there in practice, we can
still change the serialization format and even throw compression at it
if necessary.

There now is an internal __IP2UNIX_RULES, which contains rules in YAML
format and the IP2UNIX_RULE_FILE environment variable is used as a
fallback so that people who want to use LD_PRELOAD directly can do so as
documented.

An advantage of directly putting the rules into the environment is that
we are now less prone to race conditions that might occur between
ip2unix and the execve() to the new process that should be wrapped.

Signed-off-by: aszlig <aszlig@nix.build>