Allow to specify rules directly via command line
First of all, this introduces a backwards-compatible change, mainly that
the rule file is no longer mandatory and is now specified using the -f
argument.

The reason for this is to make it easier to quickly run the program
without the need to create a rule file and with a less verbose
specification.

I went through a few iterations when designing the command line
arguments and eventually got to one that I think fits quite well.

Originally, the separator was ":" instead of "," but given that ":" is
used in IPv6 addresses I opted for ",".

Signed-off-by: aszlig <aszlig@nix.build>