README: Reword the paragraph about socket_wrapper
After testing a whole bunch of different applications it turns out that
our implementation isn't so error-prone after all. So let's get this a
bit more straight and say it's easier to implement, which it actually
is. The more complex parts of socket_wrapper also are unrelated to the
core functionality (eg. they implement capture support).

Signed-off-by: aszlig <aszlig@nix.build>