Replay F_SETOWN_EX and F_SETSIG correctly
With this we now have all the relevant socket options that we need to at
least theoretically support all the relevant socket options of
applications out there.

I've added a test which uses signal based asynchronous sockets, so that
we have covered exactly this use case (which I believe isn't so
widespread out there).

Signed-off-by: aszlig <aszlig@nix.build>