Rename UdsmapRule to just Rule
The prototype of this very program originally was called "udsmap" but I
renamed it to "ip2unix". However, instead of calling it Ip2UnixRule or
something like that, I opted for a shorter name instead because it
shouldn't collide with anything and *if* that should be the case someday
we can still use namespaces.

Signed-off-by: aszlig <aszlig@nix.build>