Split out socket operations into separate classes
So far all of our socket handling was residing in preload.cc, along with
locking and rule handling. We already factored out rule handling to some
degree, but still all the socket operations were still there so far.

Now the goal is to have a self-contained Socket class, which should be
structured similar to eg. Python's "socket" class, so that the file
descriptor becomes implicit and we can treat the Socket object as an
abstraction around the fd.

While this doesn't introduce new functionality it will make adding more
functionality significantly easier, especially when it comes to locking.

Right now, the sources are riddled with XXX/FIXME/TODO comments, so we
still have plenty to clean up.

Signed-off-by: aszlig <aszlig@nix.build>