Refactor locking and socket registry
So far we had the g_active_sockets mutex all over the place, which makes
working on the code a bit tricky because when implementing new features
or rewriting existing ones we might miss a lock. There also was one
occurence of such a thing, where I locked something but forget to unlock
it.

Something like this now shouldn't happen anymore for two reasons: First
of all, we're using std::lock_guard and second, we only have two
occasions where we need to lock. The first one is during Socket::create,
because the file descriptor is added to the registry and the second one
is Socket::when, which is a if-then-else kind of function that acquires
the lock if the condition is true.

The Socket class also now has Socket::active, which acts as a registry
to what we had before with g_active_sockets. However the registry now is
only indirectly accessibly via Socket::create and Socket::when, where
the latter is the only way to perform operations on Socket. That way we
require fewer locks and also have a common entry point.

Signed-off-by: aszlig <aszlig@nix.build>