Refactor Socket::make_unix to handle systemd
If we're using socket activation, we have an existing file descriptor
and so far this has been handled outside of Socket::make_unix,
duplicating almost everything except the actual call to real::socket.

This is now no longer the case and if the Socket::make_unix function
gets a file descriptor passed it won't create one but rather uses the
given one.

Signed-off-by: aszlig <aszlig@nix.build>