Decouple Rule handling from Socket
So far we had lots of duplicate checks on rule and a shared pointer to
rule within the Socket object.

Now we have *one* file (preload.cc) where we do all the rule matching
and just invoke the relevant Socket functions whenever we have a match.

This also splits apart the bind/connect functions again, because there
are really some differences we need to take into account, especially
when it comes to getpeername/getsockname.

Signed-off-by: aszlig <aszlig@nix.build>