Use explicit function type signatures for DlsymFun
So far we have used decltype() to look up the function signature, which
works fine as long as we don't have a clash between the type signature
of a function definition in preload.cc.

While this makes realcalls.hh a bit more verbose, it does have the
advantage that we require fewer headers and also avoid those collisions
mentioned.

I also sorted the symbols alphabetically.

Signed-off-by: aszlig <aszlig@nix.build>