Implement caching and replay of ioctls
We only have SIOCSPGRP and FIOASYNC to take care of, other
socket-specific ioctls are purely read-only.

This should make it possible for applications to use signal-based
asynchronous socket handling to a limited degree and we should replay
those flags correctly.

I already have a test case for that, but I'll add that later as there
are a few more things missing (hence "to a limited degree").

Signed-off-by: aszlig <aszlig@nix.build>