Socket destructor: Add a note about closing the FD
I did actually use close() here, which unfortunately has caused a few
test runs out of 10000 to fail. The reason for the failures were because
Python was expecting the close() on the Socket FD to succeed and it
failed, because the destructor has already closed the file descriptor.

So this note is there just as a reminder that I won't come up with that
idea again in the future.

After running the test suite 20000 times, not a single one is failing
now.

Signed-off-by: aszlig <aszlig@nix.build>