Implement class for dynamically assigning ports
The first implementation of this was by trying to completely generate
the array of primes using a lambda within a constexpr.

However, after enough fighting against C++ it turned out to be not so
easy because we need to know the size of the array in advance and
something like std::vector can only be used in a constexpr in the
upcoming C++20 standard.

One way to implement this using this method would be to use one
constexpr to just determine the size and use another template to
generate the actual array, which effectively would call the prime
generation twice.

So I instead opted to use a small Python script to just generate the
array of primes and substitute it accordingly in dynports.cc.

Now what this is all for is to assign port numbers in calls such as
getpeername(), getsockname(), accept() and whatnot. So for getting a new
ephemeral port, we now have DynPorts::acquire(), which will return a
pseudo-random unassigned port that was *not* previously reserved using
DynPorts::reserve().

In the event that we assigned more ports than the ephemeral range, we
simply wrap around and re-assign ports in the same way. While this might
sound problematic, it really isn't because applications shouldn't
actually care about that and even *if* they care, chances are very low
that they stop working.

They would need to check check whether it was already assigned elsewhere
*and* there would need to be more than 64512 active connections.

Signed-off-by: aszlig <aszlig@nix.build>