Implement generic port assignment for Socket
This is so that we have more real world looking ports for getpeername()
and getsockname() and implements our DynPorts system.

So far the Socket was treated as "bound" whenever the std::optional
"binding" had a value. Now however, we always set a value to "binding"
whenever we actually have a peer on the other side.

For this reason, I added another boolean flag called "bound", which
indicaties whether the Socket is really bound via the bind(2) system
call.

Tested against 2000 test runs.

Signed-off-by: aszlig <aszlig@nix.build>