Implement bind to port 0
With our DynPort implementation this now becomes an easy task, so
whenever we bind to port 0 now, a random ephemeral port is acquired and
if the bind succeed it is reserved accordingly.

I also added a test that makes sure we don't accidentally get 0 as the
port in getsockname() and the socket path also gets %p replaced
accordingly.

Signed-off-by: aszlig <aszlig@nix.build>