Socket: Get rid of "parent" field
This was originally meant for child sockets to access fields of their
respective parents. However we don't have any shared state or any
information that any of these children would need from the parent
socket.

When removing this I was wondering why the tests were failing and it
turns out that I accidentally introduced a bug where instead of working
on the file descriptor returned by real::accept4(), we've accidentally
used the fd of the parent. This had the consequence that when removing
the parent field, the parent socket gets replaced, looses all of its
referrers and subsequently gets destroyed.

So let's actually work on the fd of the children :-)

Signed-off-by: aszlig <aszlig@nix.build>