socket: Implement a more accurate accept()
So far we have hardcoded 127.0.0.1, but we already have the right
address from a previous bind() call, which even has the right address
family, so we don't need to shovel IPv4 down the target application's
throat.

Apart from that, I rearranged the function a bit so that instead of the
need to do a Socket::close, we do all the error checking first and if
everything is fine, only then create the actual children Socket object.

Signed-off-by: aszlig <aszlig@nix.build>