Add a way to specify rule file data inline
This is mainly useful for our integration tests, as usually users might
want to specify these rules via the -r arguments.

So in order to avoid creating temporary files during our tests, having
such an option is really useful and pretty straightforward to implement,
because we already have functionality in place to parse rules from a
string directly.

Signed-off-by: aszlig <aszlig@nix.build>