tests: Get rid of ip2unix_check helper
Since we have a command line argument that allows us to specify the
rules file contents directly, we don't really need that nasty and
error-prone helper.

This also makes failing tests much easier to debug, as we're now just
using subprocess.run() instead.

Also, the main reason I want to get rid of that helper is because it's
really specific to checking rule files, which we only do in one test
file.

Signed-off-by: aszlig <aszlig@nix.build>