tests: Use -F flag for ip2unix() helper function
This should no longer leave us with a bunch of temporary files laying
around and it's also less error prone.

One thing however I wanted to get rid of is the @contextmanager around
ip2unix(), but that made running tests twice as slow. The reason might
be cleanup procedures in Popen, because when our helper just yields the
other context manager, we don't actually *use* it but rather return a
normal Popen object *inside* a context manager without the context
manager actually being active.

Of course, this is buggy by itself as dangling processes are not cleaned
up correctly, but until I do not find a way to do this without slowing
down tests tremendously I'm leaving it as is.

Signed-off-by: aszlig <aszlig@nix.build>