Implement generating addrs based on SO_PEERCRED
We now have IPv6 addresses that look like this:

  * fe80::3e8:0:64:0:4132
  * fe80::3e8:0:64:0:4133
  * fe80::3e8:0:64:0:4134
  * fe80::3e8:0:64:0:4135

The addresses consist of uid (0x03E8 -> 1000), gid (0x64 -> 100) and pid
(last example: 0x4135 -> 16693).

So in theory applications could use this to perform authentication,
however the main reason why we use SO_PEERCRED is because we can
actually generate IP addresses that are unique to a certain client
and/or application.

IPv4 addresses now look like this:

  * 0.0.65.168
  * 0.0.65.169
  * 0.0.65.170
  * 0.0.65.171

As IPv4 addresses can only hold 32 bit integers and pid_t is already 32
bits, IPv4 addresses are 1:1 the PID of the peer in network byte order.

Signed-off-by: aszlig <aszlig@nix.build>