preload: Don't pass pointer to real accept4()
So far we have passed the addr and addrlen pointers to the real accept
call, which in theory would sound well if it wouldn't come with a
problem of overflow, because usually applications pass a specific struct
pointer to accept() as they usually assume a certain family.

In practice however the allocated struct could be too small to hold an
AF_UNIX sockaddr and thus the application not only could get confused
but actually crash very badly.

New we always pass nullptrs whenever we run the accept4() call on an
AF_UNIX socket, because the actual setting of the struct sockaddr is
done in Socket::accept.

Signed-off-by: aszlig <aszlig@nix.build>