Add support for pytest-timeout plugin
Meson by default has a timeout of 30 seconds and after that time the
test process is killed and we don't know what's the reason why it got
killed because we didn't get a traceback.

There is a pytest-timeout plugin, which we're using now if it's
available and it does the same, but we're getting a traceback instead.

Signed-off-by: aszlig <aszlig@nix.build>