Return correct peer/local address on loopback
We introduced generating addresses based on SO_PEERCRED a while ago,
which is fine, but there are some applications that check to see whether
the remote address really is local only.

In order to cope with this, we now have a few functions to detect
whether an address is loopback and if it's the case, we use the
respective local address, that is for bind() we use the bound address
and for connect() we use the peer address.

Signed-off-by: aszlig <aszlig@nix.build>