Add reject and rejectError rule options
This is especially useful for something like the Erlang Port Mapper
Daemon (epmd), which starts up along with every node, tries to bind to
some addresses and if it fails with EADDRINUSE, it exits *successfully*.

Independent of epmd, this might also be useful for other applications
wishing to disallow binding to a specific address or port. Something
like sendto/sendmsg or recvfrom/recvmsg however is still allowed, so it
might not work for all applications out there.

With this the rules now support one additional action "reject", which
has an optional "reject_errno" field that allows to specify the exact
errno number or string. The latter is parsed by a new file
errno_list.cc, which is generated during build.

Signed-off-by: aszlig <aszlig@nix.build>