Blackhole sockets that bind to an existing path
Let's imagine you have a rule like this:

    port=1234,path=/tmp/foo-%p.sock

If some application now tries to bind to port 1234 once for IPv4 and
another time for IPv6 or any other combination even without different
address families, everything but the first bind will fail because the
address is already in use.

In order to cope with that, we now register the expanded socket path and
if there is another socket trying to bind to the same path it will
instead bind to a socket file in a temporary directory which is unlinked
as soon as the bind call was done.

Another alternative would be to bind to an abstract Unix Socket, but
that would mean that file permissions no longer apply and we expose
sockets that the user didn't want to expose.

Signed-off-by: aszlig <aszlig@nix.build>