Fix errors whin compiling without systemd
First of all, rule validation will succeed if no socket path was given
if built without systemd support. This is because when I added the
"reject" action, I the only failure path when no socket path was given
was to error out saying that no socket activation was enabled.

Also, a set of warnings occured when building the manpage, which
resulted in a faulty manual, mainly that the labels for the rule file
example now got a hole in it. So i brought back the system_backslash
attribute and made sure that systemd-specific details are always added
at the end.

Signed-off-by: aszlig <aszlig@nix.build>