Add rule option to blackhole the socket
Fortunately we already did most of the plumbing we needed to implement
this previously when we blackholed duplicated socket paths, so all we
need to do here now is to add an option to do it directly via a rule.

This is especially useful if applications try to bind to all sorts of
weird addresses without having a way to disable this within the
application itself. So now we can just use a blackhole rule to prevent
this.

Signed-off-by: aszlig <aszlig@nix.build>