Blackhole sockets if systemd fds are exhausted
The systemd file descriptors are organized in a queue and so far if that
queue runs empty we'll bail out with a fatal error.

While this works well if the application we run only uses one socket, it
doesn't work so well if it's binding to multiple addresses where we
didn't explicitly specify rules for.

To cope with this, we just blackhole all sockets if we have a matching
rule and there are no more file descriptors. In the event a user wants
to assign those to rules, it's still possible make the rule more
explicit and add more rules catching additional file descriptors.

Signed-off-by: aszlig <aszlig@nix.build>