Fix build without sd_listen_fds_with_names
Somebody has reported that on CentOS 7 the systemd version is 219 and
thus it doesn't have sd_listen_fds_with_names() and consecutively fails
to build.

In order to address this, we now have another compile-time constant
called NO_FDNAMES, which then causes the fd_name field and all the
handling around it to not be included.

I also modified the tests so that they take this into account.

Signed-off-by: aszlig <aszlig@nix.build>
Fixes: https://github.com/nixcloud/ip2unix/issues/2