meson: Move -fPIC to cflags
The reason this has worked so far with Nix is because we do apply -fPIC
by default with the additional hardening flags added by the cc-wrapper.

When hardening is disabled, we suddenly end up with a linking error,
because the object files are not position independent.

This is now fixed by moving -fPIC from ldflags to cflags, which is where
they should actually belong.

Signed-off-by: aszlig <aszlig@nix.build>