Implement wrappers for recv{from,msg}/send{to,msg}
This was one of the most annoying things to implement, because of a
situation like this illustrated in Python:

  On the server:

    >>> sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
    >>> sock.bind(('127.0.0.1', 1234))
    >>> sock.recvfrom(10)
    (b'foobarxxxx', ('127.0.0.1', 39059))

  On the client:

    >>> sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
    >>> sock.sendto(b'foobarxxxx', 0, ('127.0.0.1', 1234))
    10

Here we have only one binding, which is on the server side, but we don't
have one on the client, because the latter is done implicitly.

When converting this 1:1 to Unix sockets, without an implicit binding,
we won't get a socket path on the server and thus we can't distinguish
one peer from another.

Tools like netcat in Unix Domain socket mode do create an explicit
binding on a temporary path. However this is just because netcat is
playing nice here.

We do the same on our side, creating a binding (reusing blackholes) as
soon as the first sendto or sendmsg call is done.

Unfortunately, if we do have a client out there which doesn't do
something like this, then we won't have any way to identify the client
in datagram mode. For stream sockets this is way easier, because we can
use SO_PEERCRED to get the PID, UID and GID. This however is not
available for datagram sockets and while there is SCM_CREDENTIALS, we
can't use it because the *client* has to *explicitly* send those to the
server.

I believe however that a scenario like this is a very specific corner
case and not widely used out there, especially because such clients are
probably pretty domain specific and probably won't need ip2unix because
they're already using AF_UNIX.

Our implementation here still has some rough edges, such that the peer
addresses are not randomly generated (only the ports are) and a few
things could be more DRY.

All the connection tests with TCP are now also run for UDP and they all
succeed. In addition I went on to do some manual tests with netcat,
socat and Erlang to see if we really have some corner cases which we
don't cover.

Signed-off-by: aszlig <aszlig@nix.build>