BlackHole: Try harder to find temporary directory
So far we just did a look up on the TMPDIR environment variable and if
that has failed, we just set /tmp as the temporary directory. In theory
this should work on most GNU/Linux systems, but let's make sure we test
a few other locations.

Apart from testing a range of environment variables we still fall back
to /tmp and /var/tmp but also the current working directory. We now also
test whether the directory is writable and if it's actually really a
directory by also resolving the realname.

Signed-off-by: aszlig <aszlig@nix.build>