Use dlopen on libc instead of RTLD_NEXT
We still fall back to using RTLD_NEXT if everything else fails, but
first we'll dlopen() the absolute path of libc.so.6 and if that fails
just use the relative path and let the runtime loader figure it out.

This has the advantage that we exactly get the symbols that we want and
don't get into a deadlock, for example when using nested calls of
ip2unix.

Signed-off-by: aszlig <aszlig@nix.build>