Allow to use relative paths for rule args
Whenever a rule path is specified via -r rather than using a rules file,
allowing to use relative paths is pretty safe, because we do pass them
down to the actual application as an absolute path.

The only way this would break is if we'd allow a relative path within
the rules file format, as this gets directly passed to the preloaded
library and thus is subject to chdir() calls.

Signed-off-by: aszlig <aszlig@nix.build>