Implement wrappers for dup, dup2 and dup3
We don't want dup*() to interfere with our socket mapping, so we need to
make sure that whenever one of the dup variants are used on one of our
wrapped sockets the mapping for them is duplicated accordingly.

Along the way I found a bug which showed up with the new test for dup,
which is that we didn't close the old file descriptor when using dup2()
in Socket::make_unix(). This is now fixed as well and the dup() test
case covers this.

Signed-off-by: aszlig <aszlig@nix.build>