README: Add instructions for building from source
This took me a while because I only use NixOS systems, so in order to
verify the steps needed for other distributions, I have installed Fedora
29, Debian 9 and current Arch Linux in virtual machines and tried to get
it to compile.

For the packages we depend on, Fedora and Arch Linux already had newer
versions available, but Debian 9 didn't even have GCC 7. So for Debian I
installed GCC 7 from testing and managed to compile and run ip2unix, but
I wouldn't recommend that to users, so I documented it accordingly and
referred to using Nix.

Given that it's such a pain to install these dependencies on various
distributions, I'm closing #1 (Can you provide a binary distribution?)
nevertheless, because the issue text contains "Or provide install
steps" and I currently don't have the resources to provide builds for
various distributions, so for now these instructions should hopefully
suffice.

Signed-off-by: aszlig <aszlig@nix.build>
Fixes: https://github.com/nixcloud/ip2unix/issues/1