Bump version to 1.0.0
This marks the first major release out of the initial prototype and
we're feature-complete to the degree of what I initially had in mind for
the project.

The main reason for not going with an initial 0.X.X release here is that
we are now completely incompatible when it comes to the rule file format
and command line options.

Signed-off-by: aszlig <aszlig@nix.build>