Implement matching of port ranges
There are a few applications out there that bind to ports in a certain
range. In order to match something like this without the ability to
specify port ranges in rules would either be possible by creating rules
for every single port in the range or make a rule without a port at all,
which might be too lax.

For rule files we now have an additional "portEnd" option and for the
shorter rule specification we can simply use port=123-456 to specify a
range.

Signed-off-by: aszlig <aszlig@nix.build>