Bump version to 1.1.0
With this I also added a changelog file which is based on "Keep a
Changelog" on https://keepachangelog.com/, just so we have a more or
less standardized way to bring in a summary of changes.

While my commit messages are usually quite detailled, this might not be
the case for contributors, so having a brief summary for notable changes
makes it easier to keep track of what's going on.

Signed-off-by: aszlig <aszlig@nix.build>