Don't unlink wrong socket if bound *and* connected
This is something that should *never* happen, but in the course of
development I somehow forgot what the Socket::sockpath member variable
was for and assigned it also within Socket::connect.

We *did* however check whether the socket was bound using bind, so this
worked for most situations. However if an application has used bind()
and thereafter connect(), the sockpath was updated and the Socket::bound
variable thus was still true, leading to the removal of the
*destination* socket path.

I've added a test case just to make sure something like this won't
happen in the future and in addition, I renamed the Socket::sockpath
variable to Socket::unlink_sockpath, so it becomes clear that the value
in there is only for paths that need to be unlinked.

Signed-off-by: aszlig <aszlig@nix.build>