Get rid of superfluous Socket::bound variable
This variable is pretty much redundant and essentially a no-op, there is
only code that assigns it and it's tested only during destruction and
Socket::close for checking whether we need to unlink the sockpath.

However, with Socket::sockpath now being Socket::unlink_sockpath, it is
really *only* set if we want it to be unlinked, so the Socket::bound
variable is just redundant here.

Signed-off-by: aszlig <aszlig@nix.build>