Move meson.build into doc subdirectory
Having meson/meson.build for generating the manpage is confusing,
because one would imagine that the meson subdirectory is for
meson-specific helpers and the like.

So moving it to doc/ is IMHO the better location, even if I find it
slightly awkward to only have one file in that directory.

Signed-off-by: aszlig <aszlig@nix.build>