meson.build: Explicitly require Python 3
This is to make sure we don't ever accidentally run with Python 2 and
meson will abort with a fatal error if only Python 2 is available.

The reason is that our integration test suite and our generators are not
backwards-compatible with Python 2 and I don't think it's worthwhile to
make it so, because most distributions already provide Python 3 and we
already use C++17, so why use Python from the stone age?

Signed-off-by: aszlig <aszlig@nix.build>