Fix recognizing long options for specifying rules
Back when introducing the long options I used the 0 value arguments of
the list of struct option for debugging. Unfortunately they stayed there
and I never properly tested the long options, neither manually nor in
the automated tests.

This is now no longer the case and it's also covered by the test suite.

Signed-off-by: aszlig <aszlig@nix.build>