release.nix: Increase test timeout to 1 hour
With Nix we already have a timeout for the overall build we can set
freely, so for CI I'd rather set those timeouts on the Hydra rather than
inside the builds themselves as it makes builds less reliable when the
machine is loaded.

On the other side if we *do* have a problem with test runs being stuck,
we can be even more certain after the 1 hour timeout is being hit.

Signed-off-by: aszlig <aszlig@nix.build>