README: Improve manpage formatting
This essentially puts Asciidoctor into compatibility mode (using a
two-line title) and it allows us to use single quotes to generate
underline content.

Unfortunately it's not rendered the same on GitHub, but even when
generating the manpage using Asciidoctor, the output is now roughly
equivalent.

In addition to that I also changed the way the copyright symbol is
rendered, so that when we generate the manpage we get an ASCII "(C)".

Still, the manpage when generated with Asciidoctor still doesn't look as
good as with AsciiDoc, so my recommendation for AsciiDoc still holds
true.

Signed-off-by: aszlig <aszlig@nix.build>