SockAddr: Use ntohs() in get_port()
For big and little endian systems this doesn't matter a lot because
htons/ntohs is a no-op on big endian systems and for little endian
systems they both do the same. It also shouldn't matter even for middle
endian systems.

However nobody knows which weird architectures will be invented in the
future or are already existing.

Nevertheless this also makes it more clear on the intent of the code, so
let's use the right conversion function.

Signed-off-by: aszlig <aszlig@nix.build>