Add a new "ignore" rule flag
This is very useful if you have rules like this:

  - direction: outgoing
    port: 53
    ignore: true
  - direction: outgoing
    reject: true

The above rules still allow to connect to 53 (usually DNS) while
everything else will be rejected.

If we want to do something similar without such an option, we would need
to make ugly workarounds like this:

  - direction: outgoing
    port: 0
    portEnd: 52
    reject: true
  - direction: outgoing
    port: 54
    portEnd: 65535
    reject: true

So with the "ignore" flag we can simply say "don't ever turn a socket
matching this rule into a Unix socket".

Signed-off-by: aszlig <aszlig@nix.build>