release.nix: Generate a build status badge
I could have done that using Hydra's RunCommand plugin, but that would
involve having a specially configured Hydra just for the project.

So instead, we're querying the Hydra using its REST-API and trying to
find the badge job itself. That way we get the evaluation the badge job
is in and can wait until all of the builds are passing.

Unfortunately, we can't have conditions on whether a fixed-output
derivation has failed, because if that's the case the follow-up
derivation simply doesn't get built.

Fortunately, we can recycle an old trick I've used in nixpkgs for the
Chromium updater, which (ab)uses an MD5 collision to let the fixed-
output derivation always fail, but with different content.

Now MD5 is pretty much going to be deprecated in the near future in Nix,
so let's use an SHA1 collision instead.

The fixed-output derivation also uses the hash of all jobs as its name
so that it's rebuilt everytime one of the other jobs are different,
which is usually the case when we have a new evaluation.

I have not tested whether this works in conjunction with Hydra, so we
might need to fix a few bits and pieces after the next evaluation.

Signed-off-by: aszlig <aszlig@nix.build>