nix/badge: Purge Camo URL on badge update
GitHub uses an image proxy for caching, which sometimes causes the badge
to not being updated. So we now have a purgeUrl which is used to send a
HTTP PURGE request.

This still leaves a small timeframe where the old badge might be cached,
but as soon as the next evaluation occurs the URL is purged correctly.

Signed-off-by: aszlig <aszlig@nix.build>