Implement a basic logging facility
So far, ip2unix has been very quiet and has only printed fatal errors.

However, sometimes it's not so easy to find out what has led to the
fatal error, so we need a logging facility that allows us to print more
or less information based on a configured log level.

Right now this is purely based on an environment variable and has not
yet been integrated into the command line interface.

One thing to note here is that I started using *foo for some
std::optional values. I so far was reluctant to use this because if the
optional is std::nullopt we run into undefined behaviour. In this case
however it is useful for performance reasons when we potentially have a
lot of output to log.

Signed-off-by: aszlig <aszlig@nix.build>